Update HashMapDemo.java
I assumed since this code was a demo, that its intended purpose is to serve as an example of how different array methods work. For clarity and ease of learning, I would suggest the changes I made so that someone who is new to java (like me) and using this as a sort of tutorial, will easily be able to see the effects of the different HashMap methods as their results are printed to the console. I added some descriptions to be printed along with the values so that others learning from this demo will more easily understand what the methods do. Hope this helps :)