updated functions.py
changed whole program two to three functions are updated
play_game() is the main function, which performs following tasks :

Calls create_board() to create a 9×9 board and initializes with 0.
For each player (1 or 2), calls the random_place() function to randomly choose a location on board and mark that location with the player number, alternatively.
Print the board after each move.
Evaluate the board after each move to check whether a row or column or a diagonal has the same player number. If so, displays the winner name. If after 9 moves, there are no winner then displays -1.